move all ViewController methods to view.go
