+++ Makefile
@@ -1,4 +1,3 @@
-
 default: status-monitor
 
 run:
